Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
asyncpg==0.29.0
aerich==0.7.2
python-dotenv==1.0.0
python-multipart==0.0.6
twilio==8.10.0
httpx==0.25.2
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
pydantic==2.4.2
pydantic-settings==2.0.3

